First commit. Compiles 100% with precompiled headers. Crashes at dbcstores.

+++ src/game/WorldPacket.h
@@ -89,7 +89,7 @@ class WorldPacket : public ByteBuffer
         }
 
     protected:
-        uint16 m_opcode;
+        uint32 m_opcode;
         void Compress(void* dst, uint32 *dst_size, const void* src, int src_size);
 };
 #endif
